PageForm: move fields into their own folder
