+++ front/app/components/PageForm/fields/BodyField.tsx
@@ -9,7 +9,7 @@ import Error from 'components/UI/Error';
 
 // i18n
 import { FormattedMessage } from 'utils/cl-intl';
-import messages from './messages';
+import messages from '../messages';
 
 // typings
 import { Multiloc } from 'typings';
