+++ front/app/components/PageForm/fields/PageTitleField.tsx
@@ -8,7 +8,7 @@ import Error from 'components/UI/Error';
 
 // i18n
 import { FormattedMessage } from 'utils/cl-intl';
-import messages from './messages';
+import messages from '../messages';
 
 // typings
 import { Multiloc } from 'typings';
